+++ src/snapred/backend/recipe/GroupDiffCalRecipe.py
