+++ src/snapred/backend/recipe/PixelDiffCalRecipe.py
