+++ src/snapred/backend/recipe/algorithm/__init__.py
@@ -44,16 +44,6 @@ for x in all_module_names:
         # but it also is dead code and never used
         continue
 
-    if x == "PixelDiffractionCalibration":
-        # this is actually a recipe being temporarily stored here
-        # move it to recipe folder and delete this once review passes
-        continue
-
-    if x == "GroupDiffractionCalibration":
-        # this is actually a recipe being temporarily stored here
-        # move it to recipe folder and delete this once review passes
-        continue
-
     loadModule(x)
 
 # cleanup
